Started implementing saving of assets
